Move some logic to ammonite.interp.script.SingleScriptCompiler
